관심 기능 추가 및 스웨거 문서 변경

See merge request s11-bigdata-dist-sub1/S11P21D108!3

+++ total-repo/server/src/main/java/com/d108/project/domain/favorite/favoriteBusinessArea/service/FavoriteBusinessAreaServiceImpl.java
@@ -1,40 +1,60 @@
 package com.d108.project.domain.favorite.favoriteBusinessArea.service;
 
+import com.d108.project.domain.businessArea.dto.BusinessAreaListDto;
+import com.d108.project.domain.favorite.favoriteBusinessArea.dto.FavoriteBusinessAreaRequestDto;
 import com.d108.project.domain.favorite.favoriteBusinessArea.entity.FavoriteBusinessArea;
 import com.d108.project.domain.favorite.favoriteBusinessArea.repository.FavoriteBusinessAreaRepository;
 import com.d108.project.domain.businessArea.entity.BusinessArea;
 import com.d108.project.domain.businessArea.repository.BusinessAreaRepository;
-import com.d108.project.domain.favorite.favoriteBusinessArea.service.FavoriteBusinessAreaService;
 import com.d108.project.domain.member.entity.Member;
+import com.d108.project.domain.member.repository.MemberRepository;
 import lombok.RequiredArgsConstructor;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import java.util.ArrayList;
 import java.util.List;
 
 @Service
 @RequiredArgsConstructor
 public class FavoriteBusinessAreaServiceImpl implements FavoriteBusinessAreaService {
 
-    private FavoriteBusinessAreaRepository favoriteBusinessAreaRepository;
-    private BusinessAreaRepository businessAreaRepository;
+    private final FavoriteBusinessAreaRepository favoriteBusinessAreaRepository;
+    private final BusinessAreaRepository businessAreaRepository;
+
+    private final MemberRepository memberRepository;
 
     @Override
-    public FavoriteBusinessArea createFavoriteBusinessArea(Member member, Long businessAreaId) {
-        BusinessArea businessArea = businessAreaRepository.findById(businessAreaId).orElseThrow();
+    public BusinessAreaListDto getFavoriteBusinessAreasByMember(Long memberId) {
+        List<BusinessArea> businessAreas = new ArrayList<>();
 
-        FavoriteBusinessArea favorite = FavoriteBusinessArea.toFavoriteBusinessArea(member, businessArea);
-        return favoriteBusinessAreaRepository.save(favorite);
+        favoriteBusinessAreaRepository.findAllByMemberId(memberId)
+                .forEach(favoriteBusinessArea -> businessAreas.add(favoriteBusinessArea.getBusinessArea()));
+
+        return BusinessAreaListDto.createBusinessAreaListDto(businessAreas);
     }
 
     @Override
-    public void deleteFavoriteBusinessArea(Long id) {
+    public void createFavoriteBusinessArea(Long memberId, FavoriteBusinessAreaRequestDto favoriteBusinessAreaRequestDto) {
+        Member member = memberRepository.getReferenceById(memberId);
+        BusinessArea businessArea = businessAreaRepository.findById(favoriteBusinessAreaRequestDto.businessAreaId()).orElseThrow();
+
+        FavoriteBusinessArea favorite = FavoriteBusinessArea.toFavoriteBusinessArea(member, businessArea);
+        favoriteBusinessAreaRepository.save(favorite);
+    }
+
+    @Override
+    public void deleteFavoriteBusinessArea(Long memberId, Long id) {
+        FavoriteBusinessArea favoriteBusinessArea = favoriteBusinessAreaRepository.findById(id)
+                .orElseThrow(() -> new IllegalArgumentException("즐겨찾기가 존재하지 않습니다."));
+
+        Member member = favoriteBusinessArea.getMember();
+
+        if (!member.getId().equals(memberId)) {
+            throw new IllegalArgumentException("즐겨찾기를 삭제할 수 없습니다.");
+        }
+
         favoriteBusinessAreaRepository.deleteById(id);
     }
 
-    @Override
-    public List<Long> getFavoriteBusinessAreasByMember(Member member) {
-        return favoriteBusinessAreaRepository.findAllByMember(member)
-                .stream().map(FavoriteBusinessArea::getId).toList();
-    }
+
 }
